resolving code review
